requirements.txt
Django==4.1.5
djangorestframework==3.14.0
requests==2.26.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
